ps/2 elantech: Initial cleanups and fixes to support input devices
- Add references for driver implementation.
  References from FreeBSD 14.3, OpenBSD 7.8, Linux 6.17 sources
  and Linux 4.16 documentation.
- Add pragma marks for easy navigation
- Add missing define to fetch firware info from hardware
- Reduce size of version field
- Update probe function to be able to enable supported devices
- Update open function to add missing v4 devices
- Fix enable_absolute_mode and use same snooze time as BSDs and Linux drivers
- Fix get_resolution_v4 and set sane defaults instead of failing with error
- Update license for ps2_elantech.{h,cpp}

Reviewed-on: https://review.haiku-os.org/c/haiku/+/10011

+++ src/add-ons/kernel/bus_managers/ps2/ps2_elantech.h
@@ -1,13 +1,20 @@
 /*
- * Copyright 2013, Haiku, Inc.
+ * Copyright 2013-2025, Haiku, Inc.
  * Distributed under the terms of the MIT License.
  *
- * The elantech_model_info struct and all the hardware specs are taken from the
- * linux driver, thanks a lot!
+ * Hardware specs taken from the linux and BSDs drivers, thanks a lot!
+ *
+ * References:
+ *	- https://cgit.freebsd.org/src/tree/sys/dev/atkbdc/psm.c?h=releng/14.3
+ *	- https://cvsweb.openbsd.org/cgi-bin/cvsweb/src/sys/dev/pckbc/?only_with_tag=OPENBSD_7_8_BASE
+ *	- https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/drivers/input/mouse/elantech.c?h=v6.17
+ *	- https://www.kernel.org/doc/html/v4.16/input/devices/elantech.html
  *
  * Authors:
  *		Jérôme Duval <korli@users.berlios.de>
+ *		Samuel Rodríguez Pérez <samuelrp84@gmail.com>
  */
+
 #ifndef _PS2_ELANTECH_H
 #define _PS2_ELANTECH_H
 
@@ -25,7 +32,7 @@ typedef struct {
 	struct	packet_buffer*		ring_buffer;
 			size_t				packet_index;
 
-			uint32				version;
+			uint8				version;
 			uint32				fwVersion;
 
 			uint32				x;
